+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix: `pnpmSetup` task is not compatible with configuration cache when `preferPnpmOnPath = true`
+
 ## 0.0.2
 
 - Configure DNS entry for Gradle verification